move cooldowns into own class files
