+++ src/index.ts
@@ -14,14 +14,12 @@ const letmein = require('./js/letmein')
 
 const levels = require('../lib/levels.json');
 
+import cooldown from "./events/messages/cooldown";
+
 import { YouTube } from 'popyt';
-import leveling from './js/leveling';
+
 const yt = new YouTube(process.env.YT_TOKEN)
 
-let cooldown:stringKeyArray = [], cooldownXP:stringKeyArray = [];
-interface stringKeyArray {
-	[index:string]: any;
-}
 
 fs.readdir('./build/commands/', { withFileTypes:true }, (error, f) => {
     if (error) return console.error(error);
@@ -94,22 +92,7 @@ bot.on('message', async (msg:Discord.Message) => {
     if(msg.channel.type != "text")return;
     if(!msg.guild.available)return;
 
-    if(!cooldown[msg.author.id]) {
-        cooldown[msg.author.id] = 1;
-        setTimeout(async () => { delete cooldown[msg.author.id] }, 2500)
-    } else
-        cooldown[msg.author.id]++;
-
-    if(cooldown[msg.author.id] == 4)
-        return await msg.reply({"embed": { "title": "**Please calm down, or I'll mute you.**", "color": 13632027 }})
-    else if(cooldown[msg.author.id] == 6) {
-        await msg.member.roles.add('636254696880734238')
-        let msgReply = await msg.reply({"embed": { "title": "**You've been mute for 20 minutes. Reason : spamming.**", "color": 13632027 }})
-        setTimeout(async () => {
-            await msgReply.delete()
-            return msg.member.roles.remove('636254696880734238')
-        }, 1200000);
-    }
+    await cooldown.message(msg);
 
     if(msg.channel.id == process.env.SUGGESTIONTC) {
         await msg.react('✅');
@@ -120,23 +103,8 @@ bot.on('message', async (msg:Discord.Message) => {
     let db = mongod.db(dbName);
     let date:string = new Date().toISOString().slice(0,10)
 
-    if(!msg.content.startsWith(process.env.PREFIX)) {
-        await db.collection('stats').updateOne({ _id: date }, { $inc: { msg: 1 } }, { upsert: true })
-        if(msg.channel.id == '608630294261530624')return;
-
-        let user = await db.collection('user').findOne({ '_id': { $eq: msg.author.id } });
-
-        if(!user)
-            await db.collection('user').insertOne({ _id: msg.author.id, exp: 1, birthday: null, fc: null, hidden: false, pat: 0, hug: 0, boop: 0, slap: 0, highfive: 0 });
-        else if(!cooldownXP[msg.author.id]) {
-            await db.collection('user').updateOne({ _id: msg.author.id }, { $inc: { exp: 1 }});
-            leveling.levelCheck(msg, (user.exp+1));
-            cooldownXP[msg.author.id] = 1;
-            return setTimeout(async () => { delete cooldownXP[msg.author.id] }, 5000)
-        }
-
-        return mongod.close();
-    }
+    if(!msg.content.startsWith(process.env.PREFIX))
+        return await cooldown.exp(msg, mongod, db, date);
 
     let args = msg.content.slice(1).trim().split(/ +/g);
     let req = args.shift();
